#2816: restructure, and reimplement validator modules
